AI Chat/AI Chat Lab redux split
