+++ apps/src/aichat/redux/index.ts
@@ -4,22 +4,15 @@ export type {AichatState} from './state';
 // Only export a subset of actions that are meant to be used by components/application code
 export {
   clearChatMessages,
-  endSave,
   removeUpdateMessage,
-  resetToDefaultAiCustomizations,
-  setAiCustomizationProperty,
-  setModelCardProperty,
   setNewChatSession,
-  setShowModalType,
   setStudentChatHistory,
   setClientType,
-  setViewMode,
   addStagedFile,
   stagedFileUploadFinished,
   removeStagedFile,
   stagedFilesLimitExceeded,
   clearStagedFilesAlert,
   clearStagedFiles,
-  clearHasSetInitialCustomizations,
   setChatWorkspaceSelectedTab,
 } from './slice';
