+++ apps/src/aichat/redux/thunks/index.ts
@@ -1,13 +1,6 @@
 export {addChatEvent} from './addChatEvent';
 export {fetchUserChatHistory} from './fetchUserChatHistory';
-export {onSaveComplete} from './onSaveComplete';
-export {onSaveFail} from './onSaveFail';
-export {onSaveNoop} from './onSaveNoop';
-export {publishModelCard} from './publishModelCard';
-export {saveModelCard} from './saveModelCard';
 export {sendAnalytics} from './sendAnalytics';
 export {submitChatContents} from './submitChatContents';
 export {submitTeacherFeedback} from './submitTeacherFeedback';
-export {updateAiCustomization} from './updateAiCustomization';
-export {initializeAiCustomizations} from './initializeAiCustomizations';
 export {uploadFiles} from './uploadFiles';
