+++ apps/src/aichatLab/redux/thunks/onSaveComplete.ts
@@ -1,3 +1,4 @@
+import {sendAnalytics} from '@cdo/apps/aichat/redux';
 import {SaveType, ViewMode} from '@cdo/apps/aichatLab/types';
 import {sendProgressReport} from '@cdo/apps/code-studio/progressRedux';
 import {TestResults} from '@cdo/apps/constants';
@@ -7,8 +8,6 @@ import {AppDispatch} from '@cdo/apps/util/reduxHooks';
 
 import {endSave, setSavedAiCustomizations, setViewMode} from '../slice';
 import {findChangedProperties} from '../utils';
-
-import {sendAnalytics} from './sendAnalytics';
 
 const saveTypeToAnalyticsEvent: {[key in SaveType]: string} = {
   updateChatbot: EVENTS.UPDATE_CHATBOT,
@@ -21,7 +20,7 @@ const saveTypeToAnalyticsEvent: {[key in SaveType]: string} = {
 export const onSaveComplete =
   () => (dispatch: AppDispatch, getState: () => RootState) => {
     const {savedAiCustomizations, currentAiCustomizations, currentSaveType} =
-      getState().aichat;
+      getState().aichatLab;
 
     const changedProperties = findChangedProperties(
       savedAiCustomizations,
