+++ apps/src/aichatLab/redux/thunks/onSaveNoop.ts
@@ -9,7 +9,7 @@ export const onSaveNoop =
   () => (dispatch: AppDispatch, getState: () => RootState) => {
     // Even if no changes were saved, go to the presentation page if the user tried to publish
     // a model card.
-    if (getState().aichat.currentSaveType === 'publishModelCard') {
+    if (getState().aichatLab.currentSaveType === 'publishModelCard') {
       dispatch(setViewMode(ViewMode.PRESENTATION));
     }
     dispatch(endSave());
