+++ apps/src/aichatLab/views/modelCustomization/CompareModelsDialog.tsx
@@ -15,7 +15,7 @@ const CompareModelsDialog: React.FunctionComponent<{
 }> = ({onClose, availableModels}) => {
   const selectedModelId = useAppSelector(
     state =>
-      state.aichat.currentAiCustomizations.selectedModelId ||
+      state.aichatLab.currentAiCustomizations.selectedModelId ||
       availableModels[0].id
   );
   const chosenModelLeft = selectedModelId;
